requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
joblib==1.3.2
imbalanced-learn==0.11.0
cryptography==41.0.8
Pillow==10.1.0
opencv-python-headless==4.8.1.78
reportlab==4.0.7
python-dotenv==1.0.0
faker==20.1.0
python-dateutil==2.8.2
